Improve script snippet to play nice with fractions
These improvement prevent unwanted triggering of the script snippet.
In the examples given, | denotes cursor position before the snippet
expansion key (currently <tab>) is pressed.

Negative lookahead (?!...) prevents trigger if _ or ^ is already
followed by {. This prevents redundant/infinite wrapping, e.g. in the
case of a_{123}| which would otherwise wrap {123} -> {{123}} again.

Only wrap if there are 2 or more characters after the _ or ^. This makes
it easier to advance to the next tabstop of the surrounding snippet
rather than unnecessarily triggering the script snippet. E.g. in
\frac{a_1}{a_2|}, hitting tab will exit the fraction snippet.

The characters following the _ or ^ cannot include }. This prevents
unwanted triggering in the example \frac{a_1}{a_2|}, which otherwise
would wrap 1}{a_2 -> {1}{a_2} (thereby breaking the \frac braces).

The trigger may optionally end with }. This allows for wrapping a script
that is nested at the end, e.g.
a_123^456|  -->  a_{123^456}|  -->  a_{123^{456}}|
This is admittedly a corner case and will likely be removed in favor of
a change in behavior regarding when to trigger the wrapping of scripts
while typing.