Allow committing TODOs but not 'DO NOT COMMIT's
Also colorize and include filename and line number in the output.

Currently, a DO NOT COMMIT anywhere in the staged file will abort the
commit. A better solution might be to `git show :file` to only grep the
staged contents of the file, but this sacrifices filename and line number
in the output (see https://stackoverflow.com/a/41249693/11319230). It
also outputs all file contents to stdout instead of just the matching
line, forcing the use of grep's -q flag.

A possible alternative might be to `git stash`, check, and pop; but then
line numbers in the staged version might not match those of the file in
the working directory.