Rework git hooks for pre-commit framework compatibility
The pre-commit framework is incompatible with core.hooksPath:
https://github.com/pre-commit/pre-commit/issues/1198.

The git hooks workflow is therefore reworked as follows:
1. Custom "global" hooks are copied into all new repos by default on
   `git init` using init.templateDir.
2. To use pre-commit "locally" in some repo, delete the custom default
   pre-commit hook and `pre-commit install`. Custom hooks can still be
   used by adding them to the repo's `.pre-commit-config.yaml` (see the
   config added here for an example).

Hooks are symlinked into the templateDir so that each new repo gets a
copied pointer to a shared, version-controlled hook. This enables
changes to the hook to be automatically and immediately reflected across
all repos. The symlinking is performed at `./install` to avoid commiting
platform/user-specific paths and keep the solution portable.

A `.pre-commit-hooks.yaml` configures this repo to house hooks used by
pre-commit. The `.pre-commit-config.yaml` includes an example of how to
use these hooks in another repo through pre-commit.