Suppress chktex warning (1): Command terminated with space
Requires appending {} or % to far too many macros when they're not
actually needed.  E.g.:
    \backmatter%
    \phantomsection{}
etc.

There are reasons why {} or % are sometimes necessary:
    https://tex.stackexchange.com/a/392171/185167

But the chktex maintainer doesn't even enable this warning, so the
default list of macros that should be ignored is not updated:
    https://tex.stackexchange.com/a/627893/185167
(See also @egreg's comment on that thread: commands that shouldn't be
terminated with a space are rare.)